chore: removed unsued file and moved Questions component
